Revert "Enemy changes"
This reverts commit 0cf056cb390882d46aa118a64e51e02e1dcaec18.
