--- Coop-RPG/Assets/Prefabs/OverworldPrefabs/PatrolCharge.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7d77470460967c84c9306f90125c5c7e
-timeCreated: 1476836035
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
