--- Coop-RPG/Assets/Scripts/Database.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: eb1c5291f5a621d41b78adf368e9c5e8
-folderAsset: yes
-timeCreated: 1476833384
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
