--- Coop-RPG/Assets/Scripts/Overworld.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ce2e10b533c2ccc4eaf494da3f0bbc7c
-folderAsset: yes
-timeCreated: 1476833931
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
