--- Coop-RPG/Assets/Scripts/Overworld/PatrolCharge_Overworld.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 92b7613ddc052b74ea3d2fc0dfceca6a
-timeCreated: 1476835038
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
