--- Coop-RPG/Assets/Sprites/EnemySprites/PatrolCharge.png.meta
@@ -1,62 +1,0 @@
-fileFormatVersion: 2
-guid: 80bf72ca85fd2bc47824b50919dfb7a8
-timeCreated: 1476835752
-licenseType: Free
-TextureImporter:
-  fileIDToRecycleName: {}
-  serializedVersion: 2
-  mipmaps:
-    mipMapMode: 0
-    enableMipMap: 1
-    linearTexture: 0
-    correctGamma: 0
-    fadeOut: 0
-    borderMipMap: 0
-    mipMapFadeDistanceStart: 1
-    mipMapFadeDistanceEnd: 3
-  bumpmap:
-    convertToNormalMap: 0
-    externalNormalMap: 0
-    heightScale: 0.25
-    normalMapFilter: 0
-  isReadable: 0
-  grayScaleToAlpha: 0
-  generateCubemap: 0
-  cubemapConvolution: 0
-  cubemapConvolutionSteps: 7
-  cubemapConvolutionExponent: 1.5
-  seamlessCubemap: 0
-  textureFormat: -3
-  maxTextureSize: 2048
-  textureSettings:
-    filterMode: 0
-    aniso: 16
-    mipBias: 0
-    wrapMode: 1
-  nPOTScale: 0
-  lightmap: 0
-  rGBM: 0
-  compressionQuality: 50
-  allowsAlphaSplitting: 0
-  spriteMode: 1
-  spriteExtrude: 1
-  spriteMeshType: 1
-  alignment: 0
-  spritePivot: {x: 0.5, y: 0.5}
-  spriteBorder: {x: 0, y: 0, z: 0, w: 0}
-  spritePixelsToUnits: 64
-  alphaIsTransparency: 1
-  spriteTessellationDetail: -1
-  textureType: 8
-  buildTargetSettings: []
-  spriteSheet:
-    serializedVersion: 2
-    sprites: []
-    outline:
-    - - {x: 0, y: 2}
-      - {x: 1.7320508, y: -1.0000001}
-      - {x: -1.7320509, y: -0.9999998}
-  spritePackingTag: 
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
